gnocco: move cblog to shared/cblog
