+++ shared/cblog/cblog.go
